Stormcloud3 | Final Commit - Targetting and Iterative Deepening   - Added Move Ordering and Iterative Deepening   - Removed done / expunged Todos at top of file - Added Target evaluation matrix and weights - A full game leading to Checkmate on depth 4 is noted at the top of the Stormcloud3 File - Added OpeningBook class, though I will be reworking that to fit with position hashes used in Stormcloud 4
This version of the algorithm is too slow, and some fundamental techniques like positionkeys and position storing are part of the issue that is making it so slow / inefficient. I will be retiring from development of Stormcloud 3. Stormcloud 4 will have a different approach to offer.

I have severely enjoyed working on this and am excited to move forward with Stormcloud 4. gg.